// src/lib/models/APIResponses.types.ts
export type Animal = "bird" | "cat" | "dog" | "rabbit" | "reptile";

export interface Pet {
  id: number;
  name: string;
  animal: Animal;
  description: string;
  breed: string;
  images: string[];
  city: string;
  state: string;
}

export interface PetAPIResponse {
  numberOfResults: number;
  startIndex: number;
  endIndex: number;
  hasNext: boolean;
  pets: Pet[];
}

export interface BreedListAPIResponse {
  animal: Animal;
  breeds: string[];
}

export interface ResultRequestParams {
  location: string;
  animal: Animal;
  breed: string;
}